+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/DeleteMessages.kt
@@ -1,9 +1,11 @@
 package dev.inmo.tgbotapi.requests
 
+import dev.inmo.kslog.common.w
 import dev.inmo.tgbotapi.abstracts.types.MessageAction
 import dev.inmo.tgbotapi.abstracts.types.MessagesAction
 import dev.inmo.tgbotapi.requests.abstracts.SimpleRequest
 import dev.inmo.tgbotapi.types.*
+import dev.inmo.tgbotapi.utils.DefaultKTgBotAPIKSLog
 import dev.inmo.tgbotapi.utils.serializers.UnitFromBooleanSerializer
 import kotlinx.serialization.*
 import kotlinx.serialization.builtins.serializer
@@ -18,8 +20,8 @@ data class DeleteMessages(
     override fun method(): String = "deleteMessages"
 
     init {
-        require(messageIds.size in deleteMessagesLimit) {
-            "Messages count for deleteMessages must be in $deleteMessagesLimit range"
+        if (messageIds.size !in deleteMessagesLimit) {
+            DefaultKTgBotAPIKSLog.w("DeleteMessages", "Messages count for deleteMessages must be in $deleteMessagesLimit range")
         }
     }