+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/chat/modify/SetChatTitle.kt
@@ -1,11 +1,12 @@
 package dev.inmo.tgbotapi.requests.chat.modify
 
+import dev.inmo.kslog.common.w
 import dev.inmo.tgbotapi.abstracts.types.ChatRequest
 import dev.inmo.tgbotapi.requests.abstracts.SimpleRequest
 import dev.inmo.tgbotapi.types.*
+import dev.inmo.tgbotapi.utils.DefaultKTgBotAPIKSLog
 import dev.inmo.tgbotapi.utils.serializers.UnitFromBooleanSerializer
 import kotlinx.serialization.*
-import kotlinx.serialization.builtins.serializer
 
 @Serializable
 data class SetChatTitle (
@@ -16,7 +17,7 @@ data class SetChatTitle (
 ): ChatRequest, SimpleRequest<Unit> {
     init {
         if (title.length !in chatTitleLength) {
-            throw IllegalArgumentException("Chat title must be in $chatTitleLength range")
+            DefaultKTgBotAPIKSLog.w("SetChatTitle", "Chat title must be in $chatTitleLength range")
         }
     }