+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/stickers/SetStickerEmojiList.kt
@@ -1,8 +1,10 @@
 package dev.inmo.tgbotapi.requests.stickers
 
+import dev.inmo.kslog.common.w
 import dev.inmo.tgbotapi.requests.abstracts.*
 import dev.inmo.tgbotapi.requests.stickers.abstracts.StickerAction
 import dev.inmo.tgbotapi.types.*
+import dev.inmo.tgbotapi.utils.DefaultKTgBotAPIKSLog
 import dev.inmo.tgbotapi.utils.serializers.UnitFromBooleanSerializer
 import kotlinx.serialization.*
 import kotlinx.serialization.builtins.serializer
@@ -19,8 +21,8 @@ data class SetStickerEmojiList (
     constructor(sticker: FileId, vararg emojis: String) : this(sticker, emojis.toList())
 
     init {
-        require(emojis.size !in emojisInStickerLimit) {
-            "Emojis size should be in range $emojisInStickerLimit, but was ${emojis.size}"
+        if (emojis.size !in emojisInStickerLimit) {
+            DefaultKTgBotAPIKSLog.w("SetStickerEmojiList", "Emojis size should be in range $emojisInStickerLimit, but was ${emojis.size}")
         }
     }